temporary changes to allow for building and running typescript derivation
The typescript derivation connector would fail since the numeric inference field
is an unknown field, and it is getting its struct definitions from the master
branch.

These temporary changes allow for running a local stack with a functioning
derivation connector. This is not intended to be merged as-is and is for testing
purposes only.