go/runtime + agent: introduce /notify/shard-failure API
The runtime invokes a new /notify/shard-failure control-plane API which
is told of shard failures that have occurred within a data-plane.

At the moment, this API verifies the data-plane token and logs the
failure, but takes no further action.

Update the taskBase.heartbeatLoop() to perform this notification if the
shard's primary loop exits with a non-cancellation error.

Issue #1666